requirements.txt
PyGithub>=2.1.1
GitPython>=3.1.40
python-dotenv>=1.0.0
pandas>=2.1.4
rich>=13.7.0 